+++ src/main/java/dh/backend/clinica/service/impl/PacienteService.java
@@ -2,6 +2,7 @@ package dh.backend.clinica.service.impl;
 
 
 import dh.backend.clinica.entity.Paciente;
+import dh.backend.clinica.exception.ResourceNotFoundException;
 import dh.backend.clinica.repository.IPacienteRepository;
 import dh.backend.clinica.service.IPacienteService;
 import org.springframework.stereotype.Service;
@@ -41,8 +42,12 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
-
+        Optional<Paciente> pacienteEncontrado = buscarPorId(id);
+        if(pacienteEncontrado.isPresent()){
+            pacienteRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Paciente no encontrado");
+        }
     }
 
     @Override
